Added diagnosis tools
Sounds a lot fancier than it is. Really just added print statements so that the pi can give us feedback as to what is happening in the arduino, since there's no way other than the physical behavior of the rover to let us know what is going on within the arduino. This is more to keep this branch updated.